+++ CHANGELOG.md
@@ -5,6 +5,7 @@ hangeLog
 2.3.2 (2026-05-15)
 ------------------
 
+- BUGFIX: fix stream graph management: handle correctly terminal Duplex stream nodes
 - UPDATE: upgrade NPM dependencies and patches
 
 2.3.1 (2026-03-27)